refactor: move api routes to single external folder
