+++ examples/app/api/autocomplete/categories/route.ts
