+++ examples/app/api/autocomplete/route.ts
