+++ examples/app/api/organizations/[id]/route.ts
