+++ examples/app/api/payment-method-registration/route.ts
